+++ src/main/java/baseball/domain/gamenumber/GameNumbers.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.gamenumber;
 
 import java.util.Arrays;
 import java.util.List;
